kernel: fixed bug at C++/Rust boundary.

+++ src/libs/rust/base/src/io/mod.rs
@@ -70,8 +70,8 @@ macro_rules! llog {
 }
 
 #[no_mangle]
-pub extern "C" fn init_rust_io(pe_id: u64, name: *const i8) {
-    init(pe_id, unsafe { util::cstr_to_str(name) });
+pub extern "C" fn init_rust_io(pe_id: u32, name: *const i8) {
+    init(pe_id as u64, unsafe { util::cstr_to_str(name) });
 }
 
 /// Initializes the I/O module
